# api/manager/swagger.yaml
basePath: /api/v1
definitions:
  model.Application:
    properties:
      bio:
        type: string
      download_rate_limit:
        type: integer
      name:
        type: string
      scheduler_clusters:
        items:
          $ref: '#/definitions/model.SchedulerCluster'
        type: array
      seed_peer_clusters:
        items:
          $ref: '#/definitions/model.SeedPeerCluster'
        type: array
      state:
        type: string
      url:
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: integer
    type: object
...
  model.JSONMap:
    additionalProperties: true
    type: object
  model.Job:
    properties:
      args:
        $ref: '#/definitions/model.JSONMap'
      bio:
        type: string
      result:
        $ref: '#/definitions/model.JSONMap'
      scheduler_clusters:
        items:
          $ref: '#/definitions/model.SchedulerCluster'
        type: array
      seed_peer_clusters:
        items:
          $ref: '#/definitions/model.SeedPeerCluster'
        type: array
      state:
        type: string
      task_id:
        type: string
      type:
        type: string
      user_id:
        type: integer
    type: object
...
  model.SchedulerCluster:
    properties:
      application_id:
        type: integer
      bio:
        type: string
      client_config:
        $ref: '#/definitions/model.JSONMap'
      config:
        $ref: '#/definitions/model.JSONMap'
      is_default:
        type: boolean
      jobs:
        items:
          $ref: '#/definitions/model.Job'
        type: array
      name:
        type: string
      scopes:
        $ref: '#/definitions/model.JSONMap'
      security_group_id:
        type: integer
      seed_peer_clusters:
        items:
          $ref: '#/definitions/model.SeedPeerCluster'
        type: array
    type: object
...
  model.SeedPeerCluster:
    properties:
      application_id:
        type: integer
      bio:
        type: string
      config:
        $ref: '#/definitions/model.JSONMap'
      is_default:
        type: boolean
      jobs:
        items:
          $ref: '#/definitions/model.Job'
        type: array
      name:
        type: string
      scheduler_clusters:
        items:
          $ref: '#/definitions/model.SchedulerCluster'
        type: array
      scopes:
        $ref: '#/definitions/model.JSONMap'
      security_group_id:
        type: integer
    type: object
  model.User:
    properties:
      avatar:
        type: string
      bio:
        type: string
      email:
        type: string
      location:
        type: string
      name:
        type: string
      phone:
        type: string
      state:
        type: string
    type: object
...
  types.CreateApplicationRequest:
    properties:
      bio:
        type: string
      download_rate_limit:
        type: integer
      name:
        type: string
      state:
        enum:
        - enable
        - disable
        type: string
      url:
        type: string
      user_id:
        type: integer
    required:
    - name
    - user_id
    type: object
...
  types.UpdateApplicationRequest:
    properties:
      bio:
        type: string
      download_rate_limit:
        type: integer
      name:
        type: string
      state:
        enum:
        - enable
        - disable
        type: string
      url:
        type: string
      user_id:
        type: integer
    required:
    - user_id
    type: object
...
host: localhost:8080
info:
  contact:
    url: https://d7y.io
  description: Dragonfly Manager Server
  license:
    name: Apache 2.0
  title: Dragonfly Manager
  version: 1.0.0
paths:
  /applications:
    get:
      consumes:
      - application/json
      description: Get Applications
      parameters:
      - default: 0
        description: current page
        in: query
        name: page
        required: true
        type: integer
      - default: 10
        description: return max item count, default 10, max 50
        in: query
        maximum: 50
        minimum: 2
        name: per_page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Application'
            type: array
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: Get Applications
      tags:
      - Application
    post:
      consumes:
      - application/json
      description: create by json config
      parameters:
      - description: Application
        in: body
        name: Application
        required: true
        schema:
          $ref: '#/definitions/types.CreateApplicationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Application'
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: Create Application
      tags:
      - Application
  /applications/{id}:
    delete:
      consumes:
      - application/json
      description: Destroy by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: Destroy Application
      tags:
      - Application
    get:
      consumes:
      - application/json
      description: Get Application by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Application'
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: Get Application
      tags:
      - Application
    patch:
      consumes:
      - application/json
      description: Update by json config
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: Application
        in: body
        name: Application
        required: true
        schema:
          $ref: '#/definitions/types.UpdateApplicationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Application'
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: Update Application
      tags:
      - Application
  /applications/{id}/scheduler-clusters/{scheduler_cluster_id}:
    delete:
      consumes:
      - application/json
      description: Delete Scheduler to Application
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: scheduler cluster id
        in: path
        name: scheduler_cluster_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: Delete Scheduler to Application
      tags:
      - Application
    put:
      consumes:
      - application/json
      description: Add Scheduler to Application
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: scheduler cluster id
        in: path
        name: scheduler_cluster_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: Add Scheduler to Application
      tags:
      - Application
  /applications/{id}/seed-peer-clusters/{seed_peer_cluster_id}:
    delete:
      consumes:
      - application/json
      description: Delete SeedPeer to Application
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: seed peer cluster id
        in: path
        name: seed_peer_cluster_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: Delete SeedPeer to Application
      tags:
      - Application
    put:
      consumes:
      - application/json
      description: Add SeedPeer to Application
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: seed peer cluster id
        in: path
        name: seed_peer_cluster_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: Add SeedPeer to Application
      tags:
      - Application
...
swagger: "2.0"